Refactor radiomics features into subpackage
